feat(system): Complete questionnaire

+++ server/src/Models/Database/Participant.cs
@@ -14,16 +14,16 @@ namespace IsThisAMood.Models.Database
         public List<Entry> Entries { get; set; }
         public string AccessToken { get; set; }
         public string AlexaAccessToken {get; set;}
-        public int Communication {get; set;}
-        public int Contextualisation {get; set;}
-        public int Decision {get; set;}
-        public int Identification {get; set;}
-        public int Recognition {get; set;}
+        public float Communication {get; set;}
+        public float Contextualisation {get; set;}
+        public float Decision {get; set;}
+        public float Identification {get; set;}
+        public float Recognition {get; set;}
 
-        public int CommunicationTwo {get; set;}
-        public int ContextualisationTwo {get; set;}
-        public int DecisionTwo {get; set;}
-        public int IdentificationTwo {get; set;}
-        public int RecognitionTwo {get; set;}
+        public float CommunicationTwo {get; set;}
+        public float ContextualisationTwo {get; set;}
+        public float DecisionTwo {get; set;}
+        public float IdentificationTwo {get; set;}
+        public float RecognitionTwo {get; set;}
     }
 }
